perf: Cache-Control headers, gzip API, WebP conversion, decoding=async
- router.c: smart Cache-Control per file type
  images/fonts: max-age=31536000 immutable
  css/js: max-age=86400
  html/json: no-cache
- api.c: gzip compress /api/portfolio when Accept-Encoding: gzip
  uses zlib deflateInit2 with gzip wrapper (windowBits=15+16)
  graceful fallback to plain JSON if compression fails
- api.h: updated signature to accept raw request string
- build.bat: added -lz linker flag for zlib
- compress_images.py: rewritten to convert all images to WebP
  removes originals after conversion
  method=6 for best compression ratio
- app.js: decoding=async on highlight and gallery img tags
- index.html: decoding=async + loading=lazy on author photo
- admin.c: fixed pre-existing sign-compare warning

+++ src/admin.c
@@ -163,7 +163,7 @@ void handle_admin_change_password(SOCKET sock, const char *body, long body_size)
     }
     
     char new_pass[128] = {0};
-    size_t to_copy = body_size < sizeof(new_pass) - 1 ? (size_t)body_size : sizeof(new_pass) - 1;
+    size_t to_copy = (size_t)body_size < sizeof(new_pass) - 1 ? (size_t)body_size : sizeof(new_pass) - 1;
     memcpy(new_pass, body, to_copy);
     
     /* Remove surrounding quotes if sent as a JSON string literal */
